Search for Filings

Search for forms by keywords, Central Index Keys, and/or Standard Industrial Classification codes, filed after 2001.

Usage

sec_search(query = NULL, entity = NULL, CIKs = NULL, SICs = NULL,
  forms = NULL, date_to = NULL, date_from = "2001-01-01",
  outFile = NULL, overwrite = TRUE, start = 0, limit = Inf,
  verbose = TRUE)

Arguments

query

A character with words or phrases to search for within the form (e.g., grape "fruit juice").

entity

A character with a name, ticker, or Central Index Key.

CIKs

A vector of Central Index Keys (e.g., c(0000317540, 0000077476)).

SICs

A vector of Standard Industrial Classification codes (e.g., 2080).

forms

A vector of forms to include (e.g., "10-K").

date_to

Latest date to include results from (e.g., "2019-01-01"); defaults to current.

date_from

Earliest date to include results from; defaults to the earliest available: "2001-01-01".

outFile

Path of a file to write results to.

overwrite

Logical; if FALSE, will read in an existing results file rather than executing the search.

start

Number of results to skip.

limit

Maximum number of results to return, in steps of 100.

verbose

Logical; if FALSE, will no display status messages.

Value

A data.frame of results.

See also

Examples

if (FALSE) {

# search for 10-K filings by beverage companies
results <- sec_search(SICs = 2080, forms = "10-K", limit = 100)
}